refactor: daily_log helper 抽出 + session-commands.ts 分割 (Epic #288 PBI-2)
- R2: completePomodoroSession / completeCustomSession / completeFreeStudySession
  の「material → category_id → upsert_daily_log + console.error」重複 3 箇所を
  session-commands/_shared.ts の upsertDailyLog helper に集約。
  エラーログ形式は旧実装と同じ "<actionName> daily_log upsert failed for
  session <sessionId>:" を保ち、監視 grep パターンの互換性を維持する。

- R1: 602 行の session-commands.ts を手法別の 5 ファイルに分割:
  - _shared.ts: upsertDailyLog helper + JoinedMethod 型
  - create-session.ts: create 系 3 アクション
  - complete-card-based.ts: completeSession / completeElaborationSession
  - complete-timer-based.ts: Pomodoro / Custom / FreeStudy の 3 完了
  - complete-rest.ts: completeRestSession
  - index.ts: 既存 import パス (@/lib/actions/session-commands) 維持のため
    public API を re-export

Server Action の振る舞い (SELECT カラム / UPDATE フィールド / meta / エラー
メッセージ / 早期 return 順序) は旧実装と完全に一致。外部 12 ファイルの
import は index.ts 経由で変更不要。

closes #290
