+++ src/lib/actions/session-commands/create-session.ts
@@ -1,0 +1,161 @@
+"use server";
+
+import {
+  createSessionSchema,
+  createRestSessionSchema,
+  extractFieldErrors,
+} from "@/lib/validations/sessions";
+import { createInterleavingSessionSchema } from "@/lib/validations/interleaving";
+import type { ActionResult } from "@/lib/validations/materials";
+import { ACTION_ERRORS } from "@/lib/constants";
+import { requireAuth } from "@/lib/actions/auth-utils";
+
+export async function createSession(
+  materialId: string,
+  methodId: string,
+): Promise<ActionResult<{ id: string }>> {
+  const parsed = createSessionSchema.safeParse({ materialId, methodId });
+  if (!parsed.success) {
+    return {
+      success: false,
+      error: ACTION_ERRORS.INVALID_INPUT,
+      fieldErrors: extractFieldErrors(parsed.error),
+    };
+  }
+
+  const { user, supabase } = await requireAuth();
+
+  // RLS に加えてアプリ層でも所有者を確認し、RLS 緩和時の誤操作を防ぐ
+  const { data: material } = await supabase
+    .from("materials")
+    .select("id")
+    .eq("id", parsed.data.materialId)
+    .eq("user_id", user.id)
+    .single();
+
+  if (!material) return { success: false, error: ACTION_ERRORS.NOT_FOUND("教材") };
+
+  const { data: session, error } = await supabase
+    .from("sessions")
+    .insert({
+      material_id: parsed.data.materialId,
+      method_id: parsed.data.methodId,
+      user_id: user.id,
+      status: "in_progress",
+    })
+    .select("id")
+    .single();
+
+  if (error) return { success: false, error: ACTION_ERRORS.CREATE_FAILED("セッション") };
+
+  return { success: true, data: { id: session.id } };
+}
+
+export async function createRestSession(
+  parentSessionId: string,
+): Promise<ActionResult<{ id: string }>> {
+  const parsed = createRestSessionSchema.safeParse({ parentSessionId });
+  if (!parsed.success) {
+    return { success: false, error: ACTION_ERRORS.INVALID_INPUT };
+  }
+
+  const { user, supabase } = await requireAuth();
+
+  // RLS に加えてアプリ層でも所有者を確認し、他ユーザーのセッションへの紐付けを防ぐ
+  const { data: parentSession } = await supabase
+    .from("sessions")
+    .select("id")
+    .eq("id", parsed.data.parentSessionId)
+    .eq("user_id", user.id)
+    .single();
+
+  if (!parentSession) return { success: false, error: ACTION_ERRORS.NOT_FOUND("セッション") };
+
+  const { data: restMethod } = await supabase
+    .from("learning_methods")
+    .select("id")
+    .eq("slug", "wakeful_rest")
+    .single();
+
+  if (!restMethod) return { success: false, error: ACTION_ERRORS.NOT_FOUND("安静タイマー手法") };
+
+  const { data: session, error } = await supabase
+    .from("sessions")
+    .insert({
+      user_id: user.id,
+      method_id: restMethod.id,
+      status: "in_progress",
+      meta: { parent_session_id: parsed.data.parentSessionId },
+    })
+    .select("id")
+    .single();
+
+  if (error) return { success: false, error: ACTION_ERRORS.CREATE_FAILED("安静セッション") };
+
+  return { success: true, data: { id: session.id } };
+}
+
+export async function createInterleavingSession(
+  materialIds: string[],
+): Promise<ActionResult<{ id: string }>> {
+  const parsed = createInterleavingSessionSchema.safeParse({ materialIds });
+  if (!parsed.success) {
+    return { success: false, error: "インターリービングには2つ以上の教材が必要です" };
+  }
+
+  const { user, supabase } = await requireAuth();
+
+  const { data: method } = await supabase
+    .from("learning_methods")
+    .select("id")
+    .eq("slug", "interleaving")
+    .single();
+
+  if (!method) return { success: false, error: ACTION_ERRORS.NOT_FOUND("インターリービング手法") };
+
+  // RLS に加えてアプリ層でも全教材の所有権を確認する
+  const { data: ownedMaterials } = await supabase
+    .from("materials")
+    .select("id")
+    .eq("user_id", user.id)
+    .in("id", parsed.data.materialIds);
+
+  if (!ownedMaterials || ownedMaterials.length !== parsed.data.materialIds.length) {
+    return { success: false, error: ACTION_ERRORS.NOT_FOUND("教材") };
+  }
+
+  // material_id = NULL で interleaving セッションを作成
+  const { data: session, error: sessionError } = await supabase
+    .from("sessions")
+    .insert({
+      user_id: user.id,
+      method_id: method.id,
+      status: "in_progress",
+    })
+    .select("id")
+    .single();
+
+  if (sessionError || !session) {
+    return { success: false, error: ACTION_ERRORS.CREATE_FAILED("セッション") };
+  }
+
+  const sessionMaterialRows = parsed.data.materialIds.map((mid) => ({
+    session_id: session.id,
+    material_id: mid,
+  }));
+
+  const { error: smError } = await supabase
+    .from("session_materials")
+    .insert(sessionMaterialRows);
+
+  if (smError) {
+    // session_materials 挿入に失敗した場合、セッションを放棄状態にする
+    await supabase
+      .from("sessions")
+      .update({ status: "abandoned" })
+      .eq("id", session.id);
+    return { success: false, error: ACTION_ERRORS.CREATE_FAILED("セッション") };
+  }
+
+  return { success: true, data: { id: session.id } };
+}
