+++ src/lib/actions/session-commands/index.ts
@@ -1,0 +1,19 @@
+// 既存の import パス `@/lib/actions/session-commands` を維持するための集約ポイント。
+// 手法別に分割した実体ファイル (create-session / complete-card-based / complete-timer-based
+// / complete-rest) の Server Action を re-export する。
+export {
+  createSession,
+  createRestSession,
+  createInterleavingSession,
+} from "./create-session";
+export {
+  completeSession,
+  completeElaborationSession,
+} from "./complete-card-based";
+export {
+  completePomodoroSession,
+  completeCustomSession,
+  completeFreeStudySession,
+} from "./complete-timer-based";
+export { completeRestSession } from "./complete-rest";
+
